+++ cmd/token-push/utils_test.go
@@ -16,9 +16,12 @@
 package main
 
 import (
+	"strings"
 	"testing"
 
+	"github.com/fermitools/managed-tokens/internal/service"
 	"github.com/spf13/viper"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetPrometheusJobName(t *testing.T) {
@@ -71,6 +74,254 @@ func TestGetPrometheusJobName(t *testing.T) {
 				if result := getPrometheusJobName(); result != test.expectedJobNameResult {
 					t.Errorf("Got wrong jobname.  Expected %s, got %s", test.expectedJobNameResult, result)
 				}
+			},
+		)
+	}
+}
+
+func TestResolveDisableNotifications(t *testing.T) {
+	servicesStringSlice := []string{"experiment1_role1", "experiment2_role1", "experiment3_role1"}
+	services := make([]service.Service, 0, len(servicesStringSlice))
+	for _, s := range servicesStringSlice {
+		services = append(services, service.NewService(s))
+	}
+
+	type testCase struct {
+		description                     string
+		serviceSlice                    []service.Service
+		viperConfig                     string
+		expectedBlockAdminNotifications bool
+		expectedNoServiceNotifications  []string
+	}
+
+	testCases := []testCase{
+		{
+			"Global is false, service-levels are all non-existent",
+			services,
+			`
+{"disableNotifications": false}
+	`,
+			false,
+			make([]string, 0, len(services)),
+		},
+		{
+			"Global is false, service-levels are all false",
+			services,
+			`
+{
+	"disableNotifications": false,
+	"experiments": {
+		"experiment1": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": false
+				}
+			}
+		},
+		"experiment2": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": false
+				}
+			}
+		},
+		"experiment3": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": false
+				}
+			}
+		}
+	}
+}
+	`,
+			false,
+			make([]string, 0, len(services)),
+		},
+		{
+			"Global is true, no values set in services",
+			services,
+			`
+{
+	"disableNotifications": true,
+	"experiments": {
+		"experiment1": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		},
+		"experiment2": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		},
+		"experiment3": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		}
+	}
+}
+			`,
+			true,
+			[]string{"experiment1_role1", "experiment2_role1", "experiment3_role1"},
+		},
+		{
+			"Global is true, values set in services that match global",
+			services,
+			`
+{
+	"disableNotifications": true,
+	"experiments": {
+		"experiment1": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": true
+				}
+			}
+		},
+		"experiment2": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": true
+				}
+			}
+		},
+		"experiment3": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": true
+				}
+			}
+		}
+	}
+}
+			`,
+			true,
+			[]string{"experiment1_role1", "experiment2_role1", "experiment3_role1"},
+		},
+		{
+			"Global is false, any service-level is true", //  Note: this is fine because service email manager won't get started, so notifications won't get sent or forwarded to admin
+			services,
+			`
+{
+	"disableNotifications": false,
+	"experiments": {
+		"experiment1": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		},
+		"experiment2": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": true
+				}
+			}
+		},
+		"experiment3": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		}
+	}
+}
+			`,
+			false,
+			[]string{"experiment2_role1"},
+		},
+		{
+			"If global is true, and service-level is false",
+			services,
+			`
+{
+	"disableNotifications": true,
+	"experiments": {
+		"experiment1": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		},
+		"experiment2": {
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": false
+				}
+			}
+		},
+		"experiment3": {
+			"roles": {
+				"role1": {
+					"randomKey": "foo"
+				}
+			}
+		}
+	}
+}
+			`,
+			false,
+			[]string{"experiment1_role1", "experiment3_role1"},
+		},
+		{
+			"Experiment-overridden service case",
+			append(services, newExperimentOverriddenService("experiment1_role1", "experiment-override")),
+			`
+{
+	"disableNotifications": false,
+	"experiments": {
+		"experiment1": {
+			"roles": {
+				"role1": {
+					"fakeKey": "foo"
+				}
+			}
+		},
+		"experiment-override": {
+			"experimentOverride": "experiment1",
+			"roles": {
+				"role1": {
+					"disableNotificationsOverride": true
+				}
+			}
+		},
+		"experiment2": {
+			"roles": {
+				"role1": {
+					"fakeKey": "foo"
+				}
+			}
+		}
+	}
+}
+			`,
+			false,
+			[]string{"experiment-override_role1"},
+		},
+	}
+
+	for _, test := range testCases {
+		t.Run(
+			test.description,
+			func(t *testing.T) {
+				viper.Reset()
+				viper.SetConfigType("json")
+				viper.ReadConfig(strings.NewReader(test.viperConfig))
+
+				blockAdminNotifications, noServiceNotifications := resolveDisableNotifications(test.serviceSlice)
+				assert.Equal(t, test.expectedBlockAdminNotifications, blockAdminNotifications)
+				assert.Equal(t, test.expectedNoServiceNotifications, noServiceNotifications)
 			},
 		)
 	}
