+++ internal/cmdUtils/cmdUtils.go
@@ -38,7 +38,6 @@ import (
 	"go.opentelemetry.io/otel/codes"
 
 	"github.com/fermitools/managed-tokens/internal/environment"
-	"github.com/fermitools/managed-tokens/internal/service"
 )
 
 var (
@@ -429,35 +428,6 @@ func GetSSHOptsFromConfig(checkServiceConfigPath string) []string {
 	sshOptsString := viper.GetString(sshOptsPath)
 	sshOpts, _ := shlex.Split(sshOptsString)
 	return sshOpts
-}
-
-// ResolveDisableNotifications checks each service's configuration to determine if notifications should be disabled.
-// It takes a slice of service objects as input and returns a boolean indicating whether admin notifications should be disabled,
-// and a slice of strings containing the names of services for which notifications should be disabled.
-func ResolveDisableNotifications(services []service.Service) (bool, []string) {
-	serviceNotificationsToDisable := make([]string, 0, len(services))
-	globalDisableNotifications := viper.GetBool("disableNotifications")
-	finalDisableAdminNotifications := globalDisableNotifications
-
-	// Check each service's override
-	for _, s := range services {
-		serviceConfigPath := "experiments." + s.Experiment() + ".roles." + s.Role()
-		disableNotificationsPath, _ := GetServiceConfigOverrideKeyOrGlobalKey(serviceConfigPath, "disableNotifications")
-		serviceDisableNotifications := viper.GetBool(disableNotificationsPath)
-
-		// If global setting is to disable notifications (true), but any one of the experiments wants to have notifications sent (false),
-		// we need to send admin notifications for that service too, so override the global setting
-		if (!serviceDisableNotifications) && globalDisableNotifications {
-			finalDisableAdminNotifications = false
-		}
-
-		// If the service wants to disable notifications, either through override or from the global setting, add it to the list
-		if serviceDisableNotifications {
-			serviceNotificationsToDisable = append(serviceNotificationsToDisable, GetServiceName(s))
-		}
-	}
-
-	return finalDisableAdminNotifications, serviceNotificationsToDisable
 }
 
 // Functions to set environment.CommandEnvironment inside worker.Config
